requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
psycopg[binary]>=3.1,<4
python-dotenv>=1.0,<2
